Add files via upload
A more cohesive version of the counting program. Program functions as intended for the most part. Program will run and accept input until it is time to quit. Once the program quits data will be written to the .txt file and can be loaded up again whenever the user decides to run the program again