Add files via upload
A more cleaned up version of the code. Test code has been removed to make the code easier to understand. Code is designed to be a simple class based object oriented program made as a joke for Cohort 6. For anyone who is viewing this from Cohort 6, I hope this code can help you better understand how to structure and organize your code in future assignments.